move readConfig to its own file
so it can be used in pull command
